% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{andrew.imputed}
\alias{andrew.imputed}
\title{Andrews Curves of defected observations}
\usage{
andrew.imputed(object, dependent)
}
\arguments{
\item{dependent}{character. specifies the variable name of the dependent
variable in the original regression problem (not the imputation problem)}

\item{t}{axis position at which to evaluate}

\item{obs}{observation vector}
}
\description{
Andrews Curves are a Fourier series upon the observations in
  data. They are a tool for detecting hidden groupings, and in this case of
  defected observations a tool for determining whether there is a clear
  structure in the remaining covariates, that may explain why a certain
  observation is likely to be defected. As it is an explorative tool, where
  the ordering of the variables determines the frequency that is affected
  respectively, it is highly recommended use various column orders.
  It may even be of use to some extent to employ Principle components.
}
\examples{
title fourier series
}
