% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrew.R
\name{andrew}
\alias{andrew}
\title{Andrew's curves}
\usage{
andrew(object, ...)
}
\arguments{
\item{object}{Object of class "imputed".}

\item{...}{Further arguments to be passed to methods.}
}
\value{
Andrew's curves
}
\description{
`andrew` is a generic function used to display Andrew's curves of the
independent variables of an object of class "imputed". The function has just one single
method. See documentation of `andrew.imputed` for further description.
}
