% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrew.R
\name{andrewcore}
\alias{andrewcore}
\title{Core function to be passed to andrew.imputed}
\usage{
andrewcore(data, t = seq(-pi, pi, length.out = 100))
}
\arguments{
\item{data}{data.frame (or matrix). Contains observations rowwise and last column is a
group indicator. The indicator is responsible for coloring of the curves.
Notably, the input format exceeds the dummy format; any integer values can
be used to indicate grouping.}

\item{t}{vector (sequence). At which the fourier series is to be evaluated.}
}
\value{
Plot of Andrew's curve. Colored according to indicator.
}
\description{
`andrew.imputed` prepares the data frame to be passed to this function.
}
\note{
t is not supplied in higher function, as Fourier repeats itself anyways.
}
\examples{
# d1 <- data.frame(x = c(1,2), y = c(3,4), ind = c(1,0))
# andrewcore(data = d1)
# d2 <- as.matrix(data.frame(x = c(1,2,3), y = c(3,4,4), ind = c(1,0,2)))
# andrewcore(data = d2)
# d3 <- data.frame(x = c(1,2), ind = c(1,0))
# andrewcore(data = d3)
}
