% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62}
\alias{BLOSUM62}
\title{Object of class phyDat containing sequences for species corresponding to tree.}
\format{matrix}
\source{
citation("Biostrings")
}
\usage{
BLOSUM62
}
\description{
Standard BLOSUM62 substitution matrix
}
\keyword{datasets}
