% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{convSiteData}
\alias{convSiteData}
\title{Get number of convergent sites}
\usage{
convSiteData(
  tree,
  phydat,
  spe1,
  spe2,
  t,
  type = c("abs", "score"),
  m = getm(tree, phydat, spe1, spe2)
)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}

\item{t}{threshold}

\item{m}{The length of each gene (Up to what is desired to be evaluated). Default is entire gene}
}
\value{
The number of potentially convergent sites
}
\description{
Get the number of potentially evolved convergent sites and print the probability that this occured by chance.
}
\examples{
\dontrun{
convSiteData(smallTree, primates, "Human", "Chimp", 5)
}
}
