% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecensored.R
\name{family_fun}
\alias{family_fun}
\title{Evaluate gamlss family functions from gamlss object}
\usage{
family_fun(object, func = c("d", "p", "q", "r"), fitdata, predictdata,
  p = NULL, q = NULL, x = NULL, n = NULL, ...)
}
\arguments{
\item{object}{gamlss fit object}

\item{func}{character. "d", "p", "q", "r" for either density, distr. function,
quantile or random generation}

\item{predictdata}{dataframe. Containing the observations for which the
parameters are predicted.}

\item{p}{scalar numeric. Probability value. If quantile function used}

\item{x, q}{scalar numeric. Quantile value. If density or probability function
used respectively}

\item{n}{scalar numeric. Number of observations. If random generator function used}

\item{...}{argumenst to be passed to the called distributional function}
}
\value{
Depending on the choice of func, the respective vector of
  (d)density-, (p)cdf.- , (q)quantile- or (r)random values is returned
}
\description{
Function that, given a "gamlss" object, evaluates the
  distribution under predicted parameters of the provided dataframe
  "predictdata".
  CAUTION: Exactly ONE of the arguments x,q,p,n MUST be specfied!
}
