% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fity.R
\name{fitfull}
\alias{fitfull}
\title{[Add title here]}
\usage{
fitfull(ymu_formula, ysigma_formula, ynu_formula, ytau_formula, yfamily,
  xmu_formula, xsigma_formula, xnu_formula, xtau_formula, xfamily, data,
  indicator, censtype, ...)
}
\arguments{
\item{...}{}
}
\value{
Fitted y-model
}
\description{
[Add title here]
}
