% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedata.R
\name{generateblankdata}
\alias{generateblankdata}
\title{Generating a blank data frame}
\usage{
generateblankdata(varnames, n, correlation = NULL)
}
\arguments{
\item{varnames}{character vector. Specifies variables to be created.}

\item{n}{integer. Number of observations.}

\item{correlation}{Symmetric correlation matix}
}
\value{
data.frame
}
\description{
This function generates a blank data.frame with uniformly
  distributed variables specified in varnames. The variables are drawn i.i.d.
  uniform distribution if correlation = NULL, and are drawn from a correlated
  uniform distribution if a correlation matrix is supplied. The algorithm for the
  introduction of correlation is follows to: \cr
  https://www.r-bloggers.com/easily-generate-correlated-variables-from-any-distribution-without-copulas/ \cr
  The idea is: covariates specified in variablenames are inversely generated
  with a surrogate Multivariate normal distribution to establish correlation
  between mvnormal draws. The draws' cumulative normal probabilities are
  evaluated in the Uniform distribution to arrive at correlated uniformly
  distributed covariates.
}
