% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionals.R
\name{getBranchLength}
\alias{getBranchLength}
\title{Find the length of a branch from a given species to a given internal node.}
\usage{
getBranchLength(tree, spe, nodeNum)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{spe}{The name of the species}

\item{nodeNum}{The number of the internal node desired.}
}
\value{
The length of the branch between nodeNum and spe as given by tree.
}
\description{
Find the length of a branch from a given species to a given internal node.
}
\examples{
getBranchLength(tree, "Mouse", 15)
}
