% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionals.R
\name{getConvergent}
\alias{getConvergent}
\title{For a given species, find others in the tree that satisfy the conditions of convergent evolution at a certain position with a less than 0.05 chance of this having occured by chance.}
\usage{
getConvergent(tree, phydat, spe, pos, anc)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat}

\item{spe}{The species to compare others in the tree to}

\item{pos}{The position at which to compare the genes}

\item{anc}{Optional: the value of the ancestral amino acid at position pos (if not supplied, will be predicted using ancestral reconstruction)}
}
\value{
A vector of species names which satisfy the conditions listed above.
getConvergent(tree, primates, "Human", 1, "L")
}
\description{
For a given species, find others in the tree that satisfy the conditions of convergent evolution at a certain position with a less than 0.05 chance of this having occured by chance.
}
