% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probByChance.R
\name{getf}
\alias{getf}
\title{Get f}
\usage{
getf(
  tree,
  phydat,
  spe1,
  spe2,
  p = (1/20),
  type = c("abs", "score"),
  threshold = NA
)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{phydat}{a phyDat object containing the relevant alignment}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}

\item{p}{Probability of occurence for any AA (default 1/20)}

\item{type}{type of analysis: absolute ("abs") or by score ("score")}

\item{threshold}{minimum threshold for score option}
}
\value{
probability (f)
}
\description{
Obtain f - probability of any configuration satisfying the hypothesis of
convergent evolution under the constraint of the given tree
}
