% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionals.R
\name{getm}
\alias{getm}
\title{Get the length of two genes. (Genes must be of equal length.)}
\usage{
getm(tree, phydat, spe1, spe2)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}
}
\value{
The length of each gene.
getm(tree, primates, "Mouse", "Bovine")
}
\description{
Get the length of two genes. (Genes must be of equal length.)
}
