% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionals.R
\name{getMostRecentCommonAncestor}
\alias{getMostRecentCommonAncestor}
\title{Get the most recent common ancestor between two nodes as given by the given tree.}
\usage{
getMostRecentCommonAncestor(tree, spe1, spe2, areLeaves = TRUE)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{areLeaves}{Whether or not spe1 and spe2 are species (false indicates both internal nodes).}

\item{spe}{The name of species 1 or the number of internal node 1}

\item{nodeNum}{The name of species 2 or the number of internal node 2}
}
\value{
The number of the internal node that is the most recent point of divergence.
}
\description{
Get the most recent common ancestor between two nodes as given by the given tree.
}
\examples{
getMostRecentCommonAncestor(tree, "Mouse", "Bovine")
}
