% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionals.R
\name{mapLetters}
\alias{mapLetters}
\title{Map DNA bases to a number vector with 1=A, 2=C, 3=G, 4=T.}
\usage{
mapLetters(numVec)
}
\arguments{
\item{numVec}{The vector to be mapped.}

\item{tree}{A phylogenetic tree}
}
\value{
A vector of equivalent length with each value mapped to the appropriate base.
}
\description{
Map DNA bases to a number vector with 1=A, 2=C, 3=G, 4=T.
}
\examples{
mapLetters(c(1,2,3,4))
}
