% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areCondSatisfied.R
\name{matrixtoAAString}
\alias{matrixtoAAString}
\title{matrixtoAAString}
\usage{
matrixtoAAString(m)
}
\arguments{
\item{m}{matrix to be converted}
}
\value{
AAString object
}
\description{
Convert matrix generated during workflow of areCondSatisfied to AA sequence.
}
