% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{pam}
\alias{pam}
\title{Standard PAM matrix.}
\format{A matrix with 20 rows and 20 columns.}
\source{
Dayhoff, M.O. et al. (1978) A model of evolutionary change in proteins. Atlast of Protein Sequence and Structure. p. 345-352
}
\usage{
pam
}
\description{
A matrix for computing probabilities of common AA mutations.
}
\keyword{datasets}
