% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.imputed}
\alias{plot.imputed}
\title{title plot imputations and their quantiles}
\usage{
\method{plot}{imputed}(object, boxes = TRUE, quantiles = FALSE)
}
\arguments{
\item{object}{imputex object.}
}
\description{
Given an imputex object, this funciton plots the imputations and
 optionally plots the approximate averaged quantiles of the valid part of the
 censored conditonal distributions, from which the proposed vectors were
 'drawn', before the proposals were aggregated to become the imputed vector.
}
