% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{primates}
\alias{primates}
\title{Object of class phyDat containing sequences for species corresponding to tree.}
\format{phyDat}
\usage{
primates
}
\description{
14 sequences with 232 character and 217 different site patterns.
The states are a c g t
}
\keyword{datasets}
