% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.imputed}
\alias{print.imputed}
\title{Printing an object of class "imputed"}
\usage{
\method{print}{imputed}(x, ...)
}
\arguments{
\item{x}{Object of class "imputed".}

\item{...}{print-specific arguments. See print() documentation.}
}
\value{
Retruns a print in the console.
}
\description{
Printing an object of class "imputed"
}
