% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutProb.R
\name{probOfChange}
\alias{probOfChange}
\title{Calculate mutation probability}
\usage{
probOfChange(pam, AA1, AA2, d)
}
\arguments{
\item{pam}{The PAM matrix given in ./data.}

\item{AA1}{The origin amino acid}

\item{AA2}{The target amino acid}

\item{d}{The branch length between the two sequences.}
}
\value{
A scaled probability of mutation.
}
\description{
Calculate the probability of one amino acid mutating into another over a given branch length.
}
\examples{
probOfChange(pam, "L", "K", 3)
}
