% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutProb.R
\name{probOfChange1PAM}
\alias{probOfChange1PAM}
\title{probOfChange1PAM}
\usage{
probOfChange1PAM(pam, AA1, AA2)
}
\arguments{
\item{pam}{The PAM matrix (given in ./data)}

\item{AA1}{The origin amino acid}

\item{AA2}{The target amino acid}
}
\value{
A probability of mutation.
probOfChange(pam, "L", "K")
}
\description{
Calculate the probability of one amino acid mutating into another over a distance of 1 PAM.
}
