% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probByChance.R
\name{probOfNSitesByChance}
\alias{probOfNSitesByChance}
\title{Calculate probability of N chance 'convergent' sites}
\usage{
probOfNSitesByChance(
  tree,
  phydat,
  spe1,
  spe2,
  m,
  n,
  p = (1/20),
  t = NA,
  type = c("abs", "score")
)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{PhyDat object containing corresponding alignment information}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}

\item{m}{The length of the genes. (Must of equal length)}

\item{n}{The number of potential convergent sites}

\item{p}{The fraction of the amino acid at the target position in the evaluated genome.
(Default is 1/20)}

\item{t}{threshold for score option}

\item{type}{Type of analysis: 'abs' for basic model or 'score' for by convergence score model}
}
\value{
Scaled likelihood that the amino acids satisfying the conditions of convergent evolution is by chance.
}
\description{
Calculate the probability that n sites between the two species will satisfy the conditions of convergent evolution by chance.
}
\examples{
\dontrun{
probOfNSitesByChance(tree, "Human", "Chimp", 20, 2)
}
}
