% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rgenesconvergedPlot}
\alias{rgenesconvergedPlot}
\title{Plot the convergent network associated with the given gene at the given position on top of the phylogenetic tree.}
\usage{
rgenesconvergedPlot(tree, phydat, spe, pos)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat}

\item{spe}{The name of the species to compare to}

\item{pos}{The position of interest}
}
\value{
None
rgenesconvergedPlot(tree, primates, "Mouse", 17)
}
\description{
Plot the convergent network associated with the given gene at the given position on top of the phylogenetic tree.
}
