% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedata.R
\name{simulateData}
\alias{simulateData}
\title{Simluate missing/censored data}
\usage{
simulateData(n, param.formula = list(mu = ~exp(x1), sigma = ~sin(x2)),
  variablenames = c("x1", "x2"), defect = ~x1 > 0.6 | 0.8 | 1/3 * x1,
  family = "NO", correlation = NULL)
}
\arguments{
\item{n}{number of generated observations}

\item{param.formula}{list}

\item{variablenames}{vector filled with characters specifying all variable
names, that are used in param.formulas. First string is the variable that is
to be censored}

\item{defect}{formula. specifies which data subset and how it is defected. is
specified as follows: ~subset condition | probability | defect. Subset
condition describes which part of the uniformly distributed covariate [0,1]
(whose name must be specified in variablenames) is selected to be exposed to
defection by a bernoulli variable with the specified probability. The
defection funciton is left at the users convenience The defect is some
function, that the user must specify. For missing specifiy NA.}

\item{family}{character. Specifies the gamlss family, from which the data is
drawn. e.g. 'NO' for a dependent variable drawn}

\item{correlation}{matrix. If a correlation/covariance matrix is provided, the
covariates specified in variablenames are inversely generated with a
surrogate Multivariate normal distribution to establish correlation between
draws. The draws' cumulative normal probabilities are evaluated in the
Uniform distribution to arrive at correlated uniformly distributed
covariates.}
}
\value{
List of Dataframes. 'truedata' and 'defected' are dataframes
  containing the dependent (generated according to the param.formula list),
  the generated covariates, and a censoring/missing 'indicator' The mere
  difference between the two Dataframes is, that 'defected' has arteficially generated
  censored/missing values according to the 'defect' specification.
}
\description{
Data generator for missing/censored data with Normal distribution
}
\examples{
defect = ~ x1 > 0.6 | 0.8 | 1/3*x1
missing = simulateData(n = 100, param.formula = list(mu = ~exp(x1), sigma = ~sin(x2)),
                       variablenames =  c('x1', 'x2'), defect = ~ x1 > 0.6 | 0.8 | NA, family = 'NO')
right = simulateData(n = 100, param.formula = list(mu = ~exp(x1), sigma = ~sin(x2)), 
                     variablenames =  c('x1', 'x2'), defect = ~ x1 > 0.6 | 0.8 | 1/3*x1, family = 'NO', 
                     correlation = matrix(0.7, ncol = 2, nrow = 2 + diag(2)* 0.3))
}
