% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{smallTree}
\alias{smallTree}
\title{Phylogenetic tree with 2 tips and 1 internal nodes. Subtree to tree.
Tip labels:
 Human, Chimp
Rooted; includes branch lengths.}
\format{Contains leaf names, edge matrix, edge lengths, and number of internal nodes.}
\usage{
smallTree
}
\description{
Phylogenetic tree with 2 tips and 1 internal nodes. Subtree to tree.
Tip labels:
 Human, Chimp
Rooted; includes branch lengths.
}
\keyword{datasets}
