% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.imputed}
\alias{summary.imputed}
\title{Summarizing an object of class "imputed"}
\usage{
\method{summary}{imputed}(object, ...)
}
\arguments{
\item{object}{Object of class "imputed"}

\item{...}{summary specific arguments. See summary() documentation}
}
\value{
Retruns a summary in the console
}
\description{
Summarizing an object of class "imputed"
}
