% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{primates (phylogenetic tree).}
\format{Contains leaf names, edge matrix, edge lengths, and number of internal nodes.}
\source{
Schliep K.P. (2011) phangorn: Phylogenetic analysis in R. Bioinformatics, 27(4) 592-593
}
\usage{
tree
}
\description{
Phylogenetic tree with 14 tips and 12 internal nodes.
Tip labels: Mouse, Bovine, Lemur, Tarsier, Squir Monk, Jpn Macaq, ...
Unrooted; includes branch lengths.
}
\keyword{datasets}
