% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrew.R
\name{andrew.imputed}
\alias{andrew.imputed}
\title{Andrew's Curves of covariates for "imputed" class objects}
\usage{
\method{andrew}{imputed}(object, dependent, ordering = NULL, ...)
}
\arguments{
\item{object}{Object of class "imputed".}

\item{dependent}{character. Name of the dependent variable in the original
(not imputation) regression problem. It is removed, as the information
contained is dubious: Covariates cause the dependent and not vice versa.}

\item{ordering}{vector of characters. Names of the covariates supplied to
imputex. The argument is optional and allows to shuffle the dataframe.
Thereby, the covariates are associated with different Fourier frequencies.
It is highly recommended to make use of this option. As syntax sugar, it is
possible to specify only the first few variables and leave the remaining
ordering in the dataframe intact.}

\item{...}{Further arguments to be passed (e.g. of the imputed object)}
}
\description{
Andrew's Curves are a Fourier series upon the observations in
  data. They are a tool for detecting hidden groupings, and in this case of
  defected observations, they are a tool for determining whether there is a clear
  structure in the remaining covariates, that may explain why a certain
  observation is likely to be defected. As it is an explorative tool, where
  the ordering of the variables determines the frequency that is associated
  respectively, it is highly recommended to use various column orders. It may
  even be of use, to some extent, to employ Principle Components. Note that
  the defected, dependent and defect-indicator (as well as the lower bound in the
  interval case) variables are not considered for the Andrew's curve, as the
  information contained is ambiguous and misleading. Particulary, the dependent
  variable of the actual regression problem (not the imputation problem) is
  misleading, as it is caused by the covariates and not vice versa. \cr
  Further note, if after deleting those columns only one covariate remains, 
  the fourier will correctly return parallel lines: each value of that
  covariate is devided by sqrt(2). This is a fourier feature, not a bug.
}
\examples{
fright <- simulateData(n= 150,
                      param.formula = list(mu = ~ exp(x1) + x2+ x3, sigma = ~sin(x2)),
                      name = 'x1', subset = ~ x1 > 0.6, prob = 0.8 , damage = 1/3,
                      family = 'NO',
                      correlation = matrix(c(1, 0.3, 0.2,
                                             0.3, 1, 0.4,
                                             0.2, 0.4, 1), nrow = 3))
                                            
d <- imputex(data = fright$defected,
             xmu_formula = x1 ~ y + x2 + x3,
             xsigma_formula = ~x2,
             xnu_formula = ~1,
             xtau_formula = ~1,
             xfamily = NO(mu.link = 'identity'),
             indicator = "indicator",
             censtype= 'right' )
             
andrew(object = d, dependent = 'y', ordering = c('x3'))
}
