% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{andrew}
\alias{andrew}
\title{Andrew's curves}
\usage{
andrew(object, dependent, ordering = NULL, ...)
}
\arguments{
\item{object}{Object of class "imputed".}

\item{dependent}{character. Specifies the variable name of the dependent
variable in the original regression problem (not the imputation problem).}

\item{ordering}{character vector, specifying the order of the variables in
the Andrew's curve. Note that the ordering relates to the frequency in a
fourier that is associated with a covariate.}

\item{...}{Further arguments to be passed.}
}
\value{
Andrew's curves
}
\description{
`andrew` is a generic function used to display Andrew's curves of the
independent variables of an object of class "imputed". The function has just one single
method. See documentation of `andrew.imputed` for further description.
}
