% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{andrew}
\alias{andrew}
\title{Andrews Curves of defected observations}
\usage{
andrew(object, dependent, ordering = NULL)
}
\arguments{
\item{object}{of class "imputed"}

\item{dependent}{character. specifies the variable name of the dependent
variable in the original regression problem (not the imputation problem)}

\item{ordering}{character vector, specifying the order of the variables in
the andrews curve. Note that the ordering relates to the frequency in a
fourier that is associated with a covariate.}
}
\description{
Andrews Curves are a Fourier series upon the observations in
  data. They are a tool for detecting hidden groupings, and in this case of
  defected observations a tool for determining whether there is a clear
  structure in the remaining covariates, that may explain why a certain
  observation is likely to be defected. As it is an explorative tool, where
  the ordering of the variables determines the frequency that is affected
  respectively, it is highly recommended use various column orders. It may
  even be of use to some extent to employ Principle components. Note, that
  the defected, dependent and defect-indicator (and lowerbound in the
  intervalcase) variables are not considered for the andrews curve, as the
  information contained is ambigous and misleading. Particulary the dependent
  variable of the actual regression problem (not the imputation problem) is
  misleading, as it is caused by the covariates and not vice versa. Further
  Note, that given after deleting those columns only one covariate remains,
  the fourier will correctly return parallel lines: each value of that
  covariate is devided by sqrt(2). This is a feature not a bug.
}
\examples{
finterval = simulateData(n= 100,
                      param.formula = list(mu = ~exp(x1)+ x2+ x3, sigma = ~sin(x2)),
                      name = 'x1', subset = ~ (x2 < 0.3& x3<0.2), prob = 0.4, damage =list(c(0.3,0.9),c(1.2,1.5)),
                      family = 'NO',
                      correlation = matrix(c(1,0.3,0.2,
                                             0.3,1, 0.4,
                                             0.2,0.4,1), nrow = 3))
d = imputex(data = finterval$defected,
            xmu_formula= x1~ y+x2+x3,
            xsigma_formula = ~x2,
            xfamily = NO(mu.link = 'identity'),
            indicator = "indicator",
            censtype= 'interval',
            intervalstart = 'lower')
andrew(d, dependent = 'y')
}
