% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrew.R
\name{andrewcore}
\alias{andrewcore}
\title{Core function to be passed to andrew.imputed}
\usage{
andrewcore(data, t = seq(-pi, pi, length.out = 100))
}
\arguments{
\item{data}{data.frame (or matrix). Contains observations rowwise and last column is a
group indicator. The indicator is responsible for coloring of the curves.
Notably, the input format exceeds the dummy format; any integer values can
be used to indicate grouping.}

\item{t}{vector (sequence). At which the fourier series is to be evaluated.}
}
\value{
Plot of Andrew's curve. Colored according to indicator.
}
\description{
`andrew.imputed` prepares the data frame to be passed to this function.
}
