% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionals.R
\name{areCondSatisfied}
\alias{areCondSatisfied}
\title{Determine if convergently evolved}
\usage{
areCondSatisfied(tree, phydat, spe1, spe2, pos, anc)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}

\item{pos}{The position at which to evaluate if conditions are satisfied}

\item{anc}{Optional: the value of the ancestral amino acid at position pos (if not supplied, will be predicted using ancestral reconstruction)}
}
\value{
TRUE if conditions are satisfied; FALSE if they are not.
}
\description{
Determine whether conditions are satisfied for two species to be convergently evolved at a given position.
}
\examples{
areCondSatisfied(tree, primates, "Mouse", "Bovine", 17)
areCondSatisfied(tree, primates, "Mouse", "Bovine", 1, "L")
}
