% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areCondSatisfied.R
\name{convertToAA}
\alias{convertToAA}
\title{convertToAA}
\usage{
convertToAA(acctranData)
}
\arguments{
\item{acctranData}{A matrix of binary data as returned by phangorn function acctran().}
}
\value{
A matrix of amino acid data.
}
\description{
Convert a binary matrix representing the genomic sequence to one representing the amino acid sequence.
}
