% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getConvergent}
\alias{getConvergent}
\title{Find convergent species}
\usage{
getConvergent(tree, phydat, spe, pos, type = c("abs, score"), t = NA)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat}

\item{spe}{The species to compare others in the tree to}

\item{pos}{The position at which to compare the genes}

\item{t}{threshold}
}
\value{
A vector of species names which satisfy the conditions listed above.
}
\description{
For a given species, find others in the tree that satisfy the conditions of convergent evolution at a certain position with a less than 0.05 chance of this having occured by chance.
}
\examples{
\dontrun{
getConvergent(tree, primates, "Human", 1, 5)
}
}
