% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeData.R
\name{getMostRecentCommonAncestor}
\alias{getMostRecentCommonAncestor}
\title{getMostRecentCommonAncestor}
\usage{
getMostRecentCommonAncestor(tree, spe1, spe2)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}
}
\value{
The number of the internal node that is the most recent point of divergence.
}
\description{
Get the most recent common ancestor between two nodes as given by the given tree.
}
\examples{
getMostRecentCommonAncestor(tree, "Mouse", "Bovine")
}
