% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond.R
\name{mapLetters}
\alias{mapLetters}
\title{Convert DNA base format}
\usage{
mapLetters(numVec)
}
\arguments{
\item{numVec}{The vector to be mapped.}
}
\value{
A vector of equivalent length with each value mapped to the appropriate base.
}
\description{
Map DNA bases to a number vector with 1=A, 2=C, 3=G, 4=T.
}
