% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.imputed}
\alias{plot.imputed}
\title{Plot proposed imputations and their conditional censored distributions' average quantiles}
\usage{
\method{plot}{imputed}(x, boxes = FALSE, ...)
}
\arguments{
\item{x}{imputex object.}

\item{boxes}{boolean. indicating, whether (2) should be displayed as a boxplot}
}
\description{
Given an imputex object, this funciton plots (1) the the
  approximate averaged quantiles of the valid part of the censored conditonal
  distributions, from which the proposed vectors were 'drawn', before the
  proposals were aggregated via median to become the imputed vector. (2) the
  actual imputations for each observation. the red dots are the median
  proposals, which are infact the final imputed vector
}
\examples{
rinterval = simulateData(n= 300,
                         param.formula = list(mu = ~exp(x1), sigma = ~sin(x2)),
                         name = 'x1', subset = ~ x1 > 0.6, prob = 0.8 ,
                         damage =list(c(0.8, 0.99), c(1.2,1.5)),
                         family = 'NO',
                         correlation = NULL)
d  <- imputex(data = rinterval$defected,
              xmu_formula= x1~y,
              indicator = "indicator",
              censtype = 'interval',
              intervalstart = 'lower')
plot.imputed(d, boxes = FALSE)
}
