% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutProb.R
\name{probOfChange}
\alias{probOfChange}
\title{probOfChange}
\usage{
probOfChange(pam, AA1, AA2, d)
}
\arguments{
\item{pam}{The PAM matrix given in ./data.}

\item{AA1}{The origin amino acid}

\item{AA2}{The target amino acid}

\item{d}{The branch length between the two sequences.}
}
\value{
A scaled probability of mutation.
}
\description{
Calculate the probability of one amino acid mutating into another over a given branch length.
Warning: Exponential runtime; test only on very small datasets.
}
\examples{
probOfChange(pam, "L", "K", 3)
}
\references{
https://cran.r-project.org/web/packages/foreach/index.html

citation("Biostrings")
}
