% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probByChance.R
\name{probOfSiteConfig}
\alias{probOfSiteConfig}
\title{Calculate likihood of chance condition satisfaction}
\usage{
probOfSiteConfig(tree, phydat, spe1, spe2, pos, p = (1/20))
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{phydat}{An object of class phydat.}

\item{spe1}{The name of species 1}

\item{spe2}{The name of species 2}

\item{pos}{The position at which to evaluate if conditions are satisfied}

\item{p}{The fraction of the amino acid at the target position in the evaluated genome. (Default is 1/20)}
}
\value{
Scaled likelihood that the amino acids satisfying the conditions of convergent evolution is by chance.
}
\description{
Calculate the probability that an amino acid site will satisfy the conditions of convergent evolution by chance.
}
\examples{
probOfSiteConfig(tree, primates, "Human", "Chimp", 1)
}
