% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecensored.R
\name{samplecensored}
\alias{samplecensored}
\title{Inverse sampling - GAMLSS}
\usage{
samplecensored(object, censtype, predictdata, fitdata, censor)
}
\arguments{
\item{object}{gamlss. Fitted model whose parameters are predicted for predictdata.}

\item{censtype}{chr. specifies the type of censoring, for which is imputed}

\item{predictdata}{dataframe. predictdata of the missing/censored observations, for which
imputations are drawn}

\item{fitdata}{dataframe. The orignal Dataset upon which gamlss was fitted}

\item{censor}{character. Name of the to be predicted (damaged) column in predictdata.
is only required if censtype is NOT 'missing'.}
}
\value{
WRITE HERE WHAT TO BE RETURNED
}
\description{
Inverse sampling of censored variables, to impute only valid observations
conditional on the respective fit
}
