% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedata.R
\name{simulateData}
\alias{simulateData}
\title{Simluate missing/censored data}
\usage{
simulateData(n, param.formula = list(mu = ~exp(x1), sigma = ~sin(x2)),
  name = "x1", subset = ~x1 > 0.6, prob = 0.8, damage = 1/3,
  family = "NO", correlation = NULL)
}
\arguments{
\item{n}{number of generated observations}

\item{name}{character. specifies variable name to be defected.}

\item{subset}{formula. states a condition ( e.g. ~x1 > 0.6) which specifies
the fraction of observations, that are to be defected. Note, that if
'subset' does not exclusivly use the 'name[d]' variable, this implies that
the independence assumption of MICE is not met (on purpose). e.g. of unmet
condition: (x2 < 0.3 & x3 < 0.2)}

\item{prob}{numeric value. Specifies the binomial probability for each
observation in 'subset' to be defected.}

\item{damage}{By users defintion, it specifies what type and how the data is
to be defected.
'damage' = NA generates missing data. 
A value between [0, 1] implies right censoring (e.g. 'damage' = 1/3), 
[1,...] left censoring. The value is used to multiply the truevalue of 
'name' in order to defect the data. 
The generalization for fixed interval factors is 'damage' = list(1/3, 4/3),
where the values specifiy the factor for the lower and the upper bound respectively.
More realistic examples can be generated with vector valued 'damage':
If 'damage' = c(0.1, 1) is a vector of length 2, it specifies the min 
and max value of a uniform distribution, from which a factor is randomly 
drawn for each observation with which the true data is multiplied.
The generalization for random interval factors is 'damage' = list(c(0.2, 1), c(1,3)),
where the first vector specifies the unif interval for factors affecting 
the lower bound and the second affecting the upper bound.
NOTE: if a list is provided, both members must either vectors or sigle values.}

\item{family}{character. Specifies the gamlss family, from which the data is
drawn. e.g. 'NO' for a dependent variable drawn}

\item{correlation}{matrix. If a correlation/covariance matrix is provided,
the drawn variables are uniformly drawn, but correlated according to this matrix.}

\item{variablenames}{vector filled with characters specifying all variable
names, that are used in param.formulas. First string is the variable that is
to be censored}
}
\value{
List of Dataframes. 'truedata' and 'defected' are dataframes
  containing the dependent (generated according to the param.formula list),
  the generated covariates, and a censoring/missing 'indicator' The mere
  difference between the two Dataframes is, that 'defected' has arteficially generated
  censored/missing values according to the 'defect' specification.
}
\description{
Data generator for missing/censored data with Normal distribution
}
