% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallTree}
\alias{smallTree}
\title{smallTree: small phylogenetic tree for testing.}
\format{Contains leaf names, edge matrix, edge lengths, and number of internal nodes.}
\source{
Edited from Schliep K.P. (2011) phangorn: Phylogenetic analysis in R. Bioinformatics, 27(4) 592-593
}
\usage{
smallTree
}
\description{
Phylogenetic tree with 2 tips and 1 internal nodes. Subtree to tree (see ?tree).
Tip labels:
 Human, Chimp
Rooted; includes branch lengths.
}
\keyword{datasets}
