% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{tree}
\alias{tree}
\title{Phylogenetic tree with 14 tips and 12 internal nodes.
Tip labels:
 Mouse, Bovine, Lemur, Tarsier, Squir Monk, Jpn Macaq, ...
Unrooted; includes branch lengths.}
\format{Contains leaf names, edge matrix, edge lengths, and number of internal nodes.}
\source{
phangorn
}
\usage{
tree
}
\description{
Phylogenetic tree with 14 tips and 12 internal nodes.
Tip labels:
 Mouse, Bovine, Lemur, Tarsier, Squir Monk, Jpn Macaq, ...
Unrooted; includes branch lengths.
}
\keyword{datasets}
